RestApp/serializers/booking_serializer.py
from rest_framework import serializers
from RestApp.models import Booking

class BookingSerializer(serializers.ModelSerializer):
    class Meta:
        model = Booking
        fields = '__all__'
        depth = 1