requirements.txt
django==5.0
djangorestframework==3.14
drf-yasg==1.21.7